requirements.txt
pandas>=2.0.0
datawrapper>=0.4.0
requests>=2.28.0
python-dotenv>=1.0.0
selenium>=4.0.0
beautifulsoup4>=4.12.0

